+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 0.1.3 (alpha)
+
+- added support for method Paladins.getplayer(player_name)
+- clients now check for expired or nil sessions and create new session after 15 minutes
+- gem is now thread safe (tests missing, so I'm not really sure about it :)
+
 ## 0.1.2 (alpha)
 
 - added support for method Paladins.ping